implemented glide, fixed network layer

+++ app/src/main/java/com/kemanci/weatherforecast/model/entity/DailyWeatherReport.kt
@@ -55,7 +55,7 @@ data class DailyWeatherReport(
     var clouds: Long,
 
     @SerializedName("pop")
-    var pop: Long,
+    var pop: Double,
 
     @SerializedName("rain")
     var rain: Double,
